Refactor code Add empty methods to be completed, these empty methods were created to ensure the program is compilable
