+++ src/main/java/brokeculator/storage/parsing/GeneralInputParser.java
@@ -1,0 +1,16 @@
+package brokeculator.storage.parsing;
+import brokeculator.command.Command;
+
+public class GeneralInputParser {
+    public GeneralInputParser() {
+        // TODO
+    }
+    public Command getCommandFromUserInput(String userInput) {
+        //TODO
+        return null;
+    }
+    public Command getCommandFromFileInput(String lineInFile) {
+        // TODO
+        return null;
+    }
+}
