+++ src/test/java/seedu/brokeculator/UITest.java
@@ -1,7 +1,7 @@
 package seedu.brokeculator;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import brokeculator.UI;
+import brokeculator.frontend.UI;
 import org.junit.jupiter.api.Test;
 public class UITest {
     @Test
